+++ codegen/masm/src/lower/fpi.rs
@@ -207,6 +207,7 @@ mod tests {
             link_info: &link_info,
             invoked: &mut invoked,
             target: Default::default(),
+            debug_frame_size: None,
             stack,
             trace_target: TraceTarget::category("codegen"),
         };